Fleshing out how it's going to work out
Still undecided on some things, such as whether to pull the button in from their generated code, and then modify that, or if I should just pull in the SDK alltogether? Trying to keep the footprint minimal, it's just a WebComponent.

I'm really not a fan of the way that the keys are just floating around out there in all that.